third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/cssom/csskeyframesrule-insertrule-blocked.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
    <meta http-equiv="Content-Security-Policy" content="style-src 'unsafe-inline' 'self'">
    <style>
        @keyframes test {
        }
    </style>
</head>
<body>
    <script>
        var rule = document.styleSheets[1].cssRules[0];
        description("CSSKeyframesRule::insertRule blocked if 'unsafe-eval' is not an allowed source of style.");
        shouldThrow("rule.insertRule('0% { top: 0; left: 0; }', 0);");
    </script>
</body>
</html>
