third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/cssom/csskeyframesrule-insertrule-blocked-expected.txt

CONSOLE ERROR: Refused to evaluate a string as CSS because 'unsafe-eval' is not an allowed source of style in the following Content Security Policy directive: "style-src 'unsafe-inline' 'self'".

CSSKeyframesRule::insertRule blocked if 'unsafe-eval' is not an allowed source of style.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rule.insertRule('0% { top: 0; left: 0; }', 0); threw exception SecurityError: Failed to execute 'insertRule' on 'CSSKeyframesRule': Refused to evaluate a string as CSS because 'unsafe-eval' is not an allowed source of style in the following Content Security Policy directive: "style-src 'unsafe-inline' 'self'".
.
PASS successfullyParsed is true

TEST COMPLETE
